% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishAddin.R
\name{publishAddin}
\alias{publishAddin}
\title{Publish reports on Atril}
\usage{
publishAddin()
}
\description{
`publishAddin()` opens an [RStudio
gadget](https://shiny.rstudio.com/articles/gadgets.html) and
[addin](http://rstudio.github.io/rstudioaddins/) that allows you to
render and publish your selected Rmd file directly to Atril.
}
